add new UseCases to Data Module

+++ Data/src/main/java/ir/omidtaheri/data/entity/GenreDataEntity.kt
@@ -1,3 +1,6 @@
 package ir.omidtaheri.data.entity
 
-data class GenreDataEntity(val id: Long, val name: String)
+data class GenreDataEntity(
+    val id: Int,
+    val name: String
+)
